fix: Remove vertical velocity threshold check
Otherwise, the projectile will not apply gravity and perform a ground collision check when the velocity is near zero, even if the object is mid air.

Initially, the "IsVericalVelocityUnderThreshold" method was used to check the velocity of the projectile after it bounced off the ground, in order to avoid too low velocities that cause bounce jittering.