feat: Add SkullItem and SkullProjectile classes with their respective factories
Skull is a ranged weapon that can kill exactly one enemy. The behavior of an object is separated by two classes representing the object in the state of an item and a projectile.

SkullItem - represents a subclass of the pickable item, and SkullProjectile - the projectile instantiated when it is used.